# src/api-docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "geek_script",
    "description": "An API for a simple e-commerce shop themed around nerdy products. The final project of [Gama Academy](https://gama.academy/) webdevelopment course. Current version is still in development.\n\nUseful links:\n  - [Backend project repository](https://github.com/tiagospeckart/geekscript-backend)\n  - [Frontend project repository](https://github.com/GabrielGameDev/geek-script-front)",
    "contact": {
      "name": "API Support",
      "email": "tiagospeckart@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "1.2.1"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Localhost for Development"
    }
  ],
...
  "paths": {
...
    "/product/admin/{id_product}": {
      "put": {
        "tags": [
          "Product",
          "Admin"
        ],
        "summary": "Updates a Product",
        "description": "Updates the data fields of a specific existing Product.",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id_product",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              },
              "example": {
                "price": 390.12
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": "ID not found"
              }
            }
          },
          "500": {
            "description": "Server side error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": "Unable to update data"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Product",
          "Admin"
        ],
        "summary": "Deletes a Product",
        "description": "Deletes a single Product.",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "id_product",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Request recieved and understood, but no need to send data back"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": "ID not found"
              }
            }
          },
          "500": {
            "description": "Unable to delete data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": "Unable to delete data"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/checkout": {
      "post": {
        "tags": [
          "Purchase"
        ],
        "summary": "Insert into system a new purchase",
        "description": "This is where purchases are confirmed. While browsing the website, User can add Products to their shopping cart. At the /checkout page, all products selected at the current session are displayed, toghether with the order total value (a sum of each item's value). When the Purchase is cofirmed, the Frontend sends a list of products Ids (as puchaseIdList) and a total purchase value. Optionally, a discount coupon name can be submitted, subtracting the order total value.",
        "operationId": "checkout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Checkout"
              },
              "example": {
                "purchaseTotal": 458.23,
                "purchaseIdList": [
                  1,
                  2
                ],
                "discountName": "coupon15"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Checkout"
                }
              }
            }
          },
          "400": {
            "description": "Unable to complete the purchase",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": "Unable to complete the purchase"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Product": {
        "required": [
          "category_id",
          "description",
          "name",
          "photo",
          "price"
        ],
        "type": "object",
        "properties": {
          "id_product": {
            "minimum": 1,
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "float"
          },
          "description": {
            "type": "string"
          },
          "category_id": {
            "minimum": 1,
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Checkout": {
        "required": [
          "purchaseIdList",
          "total"
        ],
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "format": "integer"
          },
          "purchaseIdList": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "discount_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Error": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}